Reorganize core modules/funtions
