+++ lib/elixir_sense/core/introspection.ex
@@ -8,6 +8,9 @@ defmodule ElixirSense.Core.Introspection do
   """
 
   alias Alchemist.Helpers.ModuleInfo
+  alias ElixirSense.Core.TypeInfo
+  alias ElixirSense.Core.Normalized.Typespec
+  alias ElixirSense.Core.Normalized.Code, as: NormalizedCode
 
   @type mod_fun :: {mod :: module | nil, fun :: atom | nil}
   @type markdown :: String.t
@@ -19,108 +22,6 @@ defmodule ElixirSense.Core.Introspection do
     :gen_server  => GenServer,
     :gen_event   => GenEvent
   }
-
-  @basic_types %{
-    "any" => "The top type, the set of all terms",
-    "none" => "The bottom type, contains no terms",
-    "atom" => "An atom is a constant whose name is its own value. Some other languages call these symbols",
-    "map" => "Any map",
-    "pid" => "A process identifier, pid, identifies a process",
-    "port" => "A port identifier identifies an Erlang port",
-    "reference" => "A reference is a term that is unique in an Erlang runtime system, created by calling `make_ref/0`",
-    "struct" => "Any struct",
-    "tuple" => "Tuple of any size",
-    "float" => "A floating-point number",
-    "integer" => "An integer number",
-    "neg_integer" => "A negative integer",
-    "non_neg_integer" => "A non-negative integer",
-    "pos_integer" => "A positive integer",
-    "list/1" => "Proper list ([]-terminated)",
-    "nonempty_list/1" => "Non-empty proper list",
-    "maybe_improper_list/2" => "Proper or improper list (type1=contents, type2=termination)",
-    "nonempty_improper_list/2" => "Improper list (type1=contents, type2=termination)",
-    "nonempty_maybe_improper_list/2" => "Non-empty proper or improper list"
-  }
-
-  @builtin_types %{
-    "term" => (quote do: term :: any()),
-    "arity" => (quote do: arity :: 0..255),
-    "as_boolean/1" => (quote do: as_boolean(t) :: t), # A type `t` whose value will be used as a _truthy_ value
-    "binary" => (quote do: binary :: <<_::_*8>>),
-    "bitstring" => (quote do: bitstring :: <<_::_*1>>),
-    "boolean" => (quote do: boolean :: false | true),
-    "byte" => (quote do: byte :: 0..255),
-    "char" => (quote do: char :: 0..0x10FFFF),
-    "charlist" => (quote do: charlist :: [char()]),
-    "nonempty_charlist" => (quote do: nonempty_charlist :: [char(), ...]),
-    "fun" => (quote do: fun :: (... -> any)),
-    "function" => (quote do: function :: fun()),
-    "identifier" => (quote do: identifier :: pid() | port() | reference()),
-    "iodata" => (quote do: iodata :: iolist() | binary()),
-    "iolist" => (quote do: iolist :: maybe_improper_list(byte() | binary() | iolist(), binary() | [])),
-    "keyword" => (quote do: keyword :: [{atom(), any()}]),
-    "keyword/1" => (quote do: keyword(t) :: [{atom(), t}]),
-    "list" => (quote do: list :: [any()]),
-    "nonempty_list" => (quote do: nonempty_list :: nonempty_list(any())),
-    "maybe_improper_list" => (quote do: maybe_improper_list :: maybe_improper_list(any(), any())),
-    "nonempty_maybe_improper_list" => (quote do: nonempty_maybe_improper_list :: nonempty_maybe_improper_list(any(), any())),
-    "mfa" => (quote do: mfa :: {module(), atom(), arity()}),
-    "module" => (quote do: module :: atom()),
-    "no_return" => (quote do: no_return :: none()),
-    "node" => (quote do: node :: atom()),
-    "number" => (quote do: number :: integer() | float()),
-    "struct" => (quote do: struct :: %{:__struct__ => atom(), optional(atom()) => any()}),
-    "timeout" => (quote do: timeout :: :infinity | non_neg_integer())
-  }
-
-  @doc """
-  Shim to replicate the behavior of `Code.get_docs/2` in Elixir >= 1.7
-  """
-  def get_docs(module, category) do
-    if function_exported?(Code, :fetch_docs, 1) do
-      case Code.fetch_docs(module) do
-        {:docs_v1, moduledoc_line, _beam_language, "text/markdown", moduledoc, _metadata, docs} ->
-          docs = Enum.map(docs, &to_old_format/1)
-
-          case category do
-            :moduledoc ->
-              moduledoc_en =
-                case moduledoc do
-                  %{"en" => moduledoc_en} -> moduledoc_en
-                  false -> false
-                  _ -> nil
-                end
-
-              case {moduledoc_line, moduledoc_en} do
-                {_, nil} -> nil
-                {nil, _} -> nil
-                _ -> {moduledoc_line, moduledoc_en}
-              end
-
-            :docs ->
-              Enum.filter(docs, &match?({_, _, def_type, _, _} when def_type in [:def, :defmacro], &1))
-
-            :callback_docs ->
-              Enum.filter(
-                docs,
-                &match?({_, _, kind, _} when kind in [:callback, :macrocallback], &1)
-              )
-
-            :type_docs ->
-              Enum.filter(docs, &match?({_, _, :type, _}, &1))
-
-            :all ->
-              [:moduledoc, :docs, :callback_docs, :type_docs]
-              |> Enum.map(&{&1, get_docs(module, &1)})
-          end
-
-        _ ->
-          nil
-      end
-    else
-      Module.concat([Code]).get_docs(module, category)
-    end
-  end
 
   def all_modules do
     ModuleInfo.all_applications_modules()
@@ -136,7 +37,7 @@ defmodule ElixirSense.Core.Introspection do
   end
 
   def get_signatures(mod, fun, code_docs \\ nil) do
-    docs = code_docs || get_docs(mod, :docs) || []
+    docs = code_docs || NormalizedCode.get_docs(mod, :docs) || []
     for {{f, arity}, _, _, args, text} <- docs, f == fun do
       fun_args = Enum.map(args, &format_doc_arg(&1))
       fun_str = Atom.to_string(fun)
@@ -148,7 +49,7 @@ defmodule ElixirSense.Core.Introspection do
 
   def get_func_docs_md(mod, fun) do
     docs =
-      case get_docs(mod, :docs) do
+      case NormalizedCode.get_docs(mod, :docs) do
         nil -> nil
         docs ->
           for {{f, arity}, _, _, args, text} <- docs, f == fun do
@@ -170,7 +71,7 @@ defmodule ElixirSense.Core.Introspection do
 
   def get_docs_md(mod) when is_atom(mod) do
     mod_str = module_to_string(mod)
-    case get_docs(mod, :moduledoc) do
+    case NormalizedCode.get_docs(mod, :moduledoc) do
       {_line, doc} when is_binary(doc) ->
         "> #{mod_str}\n\n" <> doc
       _ ->
@@ -211,7 +112,7 @@ defmodule ElixirSense.Core.Introspection do
     case get_callbacks_and_docs(mod) do
       {callbacks, []} ->
         Enum.map(callbacks, fn {{name, arity}, [spec | _]} ->
-          spec_ast = spec_to_quoted(name, spec)
+          spec_ast = Typespec.spec_to_quoted(name, spec)
           signature = get_typespec_signature(spec_ast, arity)
           definition = format_spec_ast(spec_ast)
           %{name: name, arity: arity, callback: "@callback #{definition}", signature: signature, doc: nil}
@@ -231,28 +132,10 @@ defmodule ElixirSense.Core.Introspection do
 
   def get_types_with_docs(module) when is_atom(module) do
     module
-    |> get_types()
+    |> Typespec.get_types()
     |> Enum.map(fn {_, {t, _, _args}} = type ->
-      %{type: format_type(type), doc: get_type_doc(module, t)}
+      %{type: format_type(type), doc: TypeInfo.get_type_doc(module, t)}
     end)
-  end
-
-  def get_types(module) when is_atom(module) do
-    if Code.ensure_loaded?(Code.Typespec) do
-      case Code.Typespec.fetch_types(module) do
-        {:ok, types} -> types
-        _            -> []
-      end
-    else
-      case old_typespec().beam_types(module) do
-        nil   -> []
-        types -> types
-      end
-    end |> reject_private_types()
-  end
-
-  defp reject_private_types(types) do
-    types |> Enum.reject(fn {type, _} -> type in [:opaque, :typep] end)
   end
 
   def extract_summary_from_docs(doc) when doc in [nil, "", false], do: ""
@@ -263,19 +146,19 @@ defmodule ElixirSense.Core.Introspection do
   end
 
   defp format_type({:opaque, type}) do
-    {:::, _, [ast, _]} = type_to_quoted(type)
+    {:::, _, [ast, _]} = Typespec.type_to_quoted(type)
     "@opaque #{format_spec_ast(ast)}"
   end
 
   defp format_type({kind, type}) do
-    ast = type_to_quoted(type)
+    ast = Typespec.type_to_quoted(type)
     "@#{kind} #{format_spec_ast(ast)}"
   end
 
   def format_spec_ast_single_line(spec_ast) do
     spec_ast
     |> Macro.prewalk(&drop_macro_env/1)
-    |> spec_ast_to_string()
+    |> TypeInfo.spec_ast_to_string()
   end
 
   def format_spec_ast(spec_ast) do
@@ -311,7 +194,7 @@ defmodule ElixirSense.Core.Introspection do
 
   def define_callback?(mod, fun, arity) do
     mod
-    |> get_callbacks()
+    |> Typespec.get_callbacks()
     |> Enum.any?(fn {{f, a}, _} -> {f, a} == {fun, arity}  end)
   end
 
@@ -331,8 +214,8 @@ defmodule ElixirSense.Core.Introspection do
           _   -> {:when, [], [return, parts.when_part]}
         end
 
-      spec     = return |> spec_ast_to_string()
-      stripped = ast |> spec_ast_to_string()
+      spec     = return |> TypeInfo.spec_ast_to_string()
+      stripped = ast |> TypeInfo.spec_ast_to_string()
       snippet  = ast |> return_to_snippet()
       %{description: stripped, spec: spec, snippet: snippet}
     end
@@ -354,49 +237,12 @@ defmodule ElixirSense.Core.Introspection do
     [ast|returns]
   end
 
-  def get_type_doc(module, type, type_n_args \\ 0) do
-    # TODO: Use `with`
-    case get_docs(module, :type_docs) do
-      nil  -> ""
-      docs ->
-        case Enum.find(docs, fn({{name, n_args}, _, _, _}) -> type == name && type_n_args == n_args end) do
-          {{_, _}, _, _, description} ->
-            description || ""
-          _ ->
-            get_builtin_type_doc(type)
-        end
-    end
-  end
-
-  def get_builtin_type_doc(type, n_args \\ 0) do
-    case @builtin_types[type_key(type, n_args)] do
-      nil -> ""
-      _ -> "Built-in type"
-    end
-  end
-
-  def get_builtin_type_spec(type, n_args \\ 0) do
-    @builtin_types[type_key(type, n_args)]
-  end
-
-  def get_basic_type_doc(type, n_args \\ 0) do
-    @basic_types[type_key(type, n_args)]
-  end
-
-  defp type_key(type, n_args) do
-    if n_args > 0 do
-      "#{type}/#{n_args}"
-    else
-      "#{type}"
-    end
-  end
-
   defp get_callback_with_doc(name, kind, doc, key, callbacks) do
     {_, [spec | _]} = List.keyfind(callbacks, key, 0)
     {_f, arity} = key
 
     spec_ast = name
-    |> spec_to_quoted(spec)
+    |> Typespec.spec_to_quoted(spec)
     |> Macro.prewalk(&drop_macro_env/1)
     signature = get_typespec_signature(spec_ast, arity)
     definition = format_spec_ast(spec_ast)
@@ -405,11 +251,11 @@ defmodule ElixirSense.Core.Introspection do
   end
 
   defp get_callbacks_and_docs(mod) do
-    callbacks = get_callbacks(mod)
+    callbacks = Typespec.get_callbacks(mod)
     docs =
       @wrapped_behaviours
       |> Map.get(mod, mod)
-      |> get_docs(:callback_docs)
+      |> NormalizedCode.get_docs(:callback_docs)
 
     {callbacks || [], docs || []}
   end
@@ -476,7 +322,7 @@ defmodule ElixirSense.Core.Introspection do
     {ast, index}
   end
   defp next_snippet(ast, index) do
-    {"${#{index}:#{spec_ast_to_string(ast)}}$", index + 1}
+    {"${#{index}:#{TypeInfo.spec_ast_to_string(ast)}}$", index + 1}
   end
 
   def param_to_var({{:=, _, [_lhs, {name, _, _} = rhs]}, arg_index}) when is_atom(name) do
@@ -524,7 +370,7 @@ defmodule ElixirSense.Core.Introspection do
     do: {:"arg#{i}", [], nil}
 
   def get_module_docs_summary(module) do
-    case get_docs module, :moduledoc do
+    case NormalizedCode.get_docs module, :moduledoc do
       {_, doc} -> extract_summary_from_docs(doc)
       _ -> ""
     end
@@ -566,31 +412,8 @@ defmodule ElixirSense.Core.Introspection do
     {"", ""}
   end
 
-  def get_module_specs(module) do
-    case beam_specs(module) do
-      nil   -> %{}
-      specs ->
-        for {_kind, {{f, a}, _spec}} = spec <- specs, into: %{} do
-          {{f, a}, spec}
-        end
-    end
-  end
-
-  def get_spec(module, function, arity) when is_atom(module) and is_atom(function) and is_integer(arity) do
-    module
-    |> get_module_specs()
-    |> Map.get({function, arity})
-  end
-
   def get_spec_as_string(module, function, arity) do
-    get_spec(module, function, arity) |> spec_to_string()
-  end
-
-  def get_function_specs(module, function) when is_atom(module) and is_atom(function) do
-    specs = module |> get_module_specs()
-    for {{f, _}, spec} <- specs, f == function do
-      spec
-    end
+    TypeInfo.get_spec(module, function, arity) |> spec_to_string()
   end
 
   def get_spec_text(mod, fun, arity) do
@@ -634,8 +457,8 @@ defmodule ElixirSense.Core.Introspection do
   end
 
   def module_functions_info(module) do
-    docs = get_docs(module, :docs) || []
-    specs = get_module_specs(module)
+    docs = NormalizedCode.get_docs(module, :docs) || []
+    specs = TypeInfo.get_module_specs(module)
     for {{f, a}, _line, func_kind, _sign, doc} = func_doc <- docs, doc != false, into: %{} do
       spec = Map.get(specs, {f, a})
       {fun_args, desc} = extract_fun_args_and_desc(func_doc)
@@ -645,10 +468,10 @@ defmodule ElixirSense.Core.Introspection do
 
   def get_callback_ast(module, callback, arity) do
     {{name, _}, [spec | _]} = module
-      |> get_callbacks()
+      |> Typespec.get_callbacks()
       |> Enum.find(fn {{f, a}, _} -> {f, a} == {callback, arity}  end)
 
-    spec_to_quoted(name, spec)
+      Typespec.spec_to_quoted(name, spec)
   end
 
   defp format_doc_arg({:\\, _, [left, right]}) do
@@ -659,10 +482,6 @@ defmodule ElixirSense.Core.Introspection do
     Atom.to_string(var)
   end
 
-  def spec_ast_to_string(ast) do
-    ast |> Macro.to_string |> String.replace("()", "")
-  end
-
   def spec_to_string(nil) do
     ""
   end
@@ -670,29 +489,10 @@ defmodule ElixirSense.Core.Introspection do
   def spec_to_string({kind, {{name, _arity}, specs}}) do
     specs
     |> Enum.map(fn spec ->
-      binary = Macro.to_string(spec_to_quoted(name, spec))
+      binary = Macro.to_string(Typespec.spec_to_quoted(name, spec))
       "@#{kind} #{binary}" |> String.replace("()", "")
     end)
     |> Enum.join("\n")
-  end
-
-  defp beam_specs(module) do
-    specs =
-      if Code.ensure_loaded?(Code.Typespec) do
-        case Code.Typespec.fetch_specs(module) do
-          {:ok, specs} -> specs
-          _            -> []
-        end
-      else
-        old_typespec().beam_specs(module)
-      end
-
-    beam_specs_tag(specs, :spec)
-  end
-
-  defp beam_specs_tag(nil, _), do: nil
-  defp beam_specs_tag(specs, tag) do
-    Enum.map(specs, &{tag, &1})
   end
 
   def actual_mod_fun(mod_fun, imports, aliases, current_module) do
@@ -770,75 +570,6 @@ defmodule ElixirSense.Core.Introspection do
   end
   defp elixir_module?(_) do
     false
-  end
-
-  defp to_old_format({{kind, name, arity}, line, signatures, docs, _meta}) do
-    docs_en =
-      case docs do
-        %{"en" => docs_en} -> docs_en
-        false -> false
-        _ -> nil
-      end
-
-    case kind do
-      kind when kind in [:function, :macro] ->
-        args_quoted =
-          signatures
-          |> Enum.join(" ")
-          |> Code.string_to_quoted()
-          |> case do
-            {:ok, {^name, _, args}} -> args
-            _ -> []
-          end
-
-        def_type =
-          case kind do
-            :function -> :def
-            :macro -> :defmacro
-          end
-
-        {{name, arity}, line, def_type, args_quoted, docs_en}
-
-      _ ->
-        {{name, arity}, line, kind, docs_en}
-    end
-  end
-
-  defp get_callbacks(mod) do
-    if Code.ensure_loaded?(Code.Typespec) do
-      case Code.Typespec.fetch_callbacks(mod) do
-        {:ok, callbacks} -> callbacks
-        _ -> []
-      end
-    else
-      old_typespec().beam_callbacks(mod)
-    end
-  end
-
-  # Workaround because Code.Typespec.typespec_to_quoted/1 is private
-  def typespec_to_quoted(type) do
-    {:::, [], [_, quoted]} = type_to_quoted({:fake_var, type, []})
-    quoted
-  end
-
-  def type_to_quoted(type) do
-    if Code.ensure_loaded?(Code.Typespec) do
-      Code.Typespec.type_to_quoted(type)
-    else
-      old_typespec().type_to_ast(type)
-    end
-  end
-
-  def spec_to_quoted(name, spec) do
-    if Code.ensure_loaded?(Code.Typespec) do
-      Code.Typespec.spec_to_quoted(name, spec)
-    else
-      old_typespec().spec_to_ast(name, spec)
-    end
-  end
-
-  defp old_typespec() do
-    Kernel.Typespec
   end
 
 end
