+++ test/elixir_sense/core/introspection_test.exs
@@ -2,7 +2,37 @@ defmodule ElixirSense.Core.IntrospectionTest do
 
   use ExUnit.Case
 
+  alias ElixirSense.Core.NormalizedTypespec
   import ElixirSense.Core.Introspection
+
+  test "format_spec_ast with one return option does not aplit the returns" do
+    type_ast = get_type_ast(GenServer, :debug)
+
+    assert format_spec_ast(type_ast) == """
+    debug :: [:trace | :log | :statistics | {:log_to_file, Path.t}]
+    """
+  end
+
+  test "format_spec_ast with more than one return option aplits the returns" do
+    type_ast = get_type_ast(GenServer, :on_start)
+
+    assert format_spec_ast(type_ast) == """
+    on_start ::
+      {:ok, pid} |
+      :ignore |
+      {:error, {:already_started, pid} | term}
+    """
+  end
+
+  test "format_spec_ast for callback" do
+    ast = get_callback_ast(GenServer, :code_change, 3)
+    assert format_spec_ast(ast) == """
+    code_change(old_vsn, state :: term, extra :: term) ::
+      {:ok, new_state :: term} |
+      {:error, reason :: term} |
+      {:down, term} when old_vsn: term
+    """
+  end
 
   test "get_callbacks_with_docs for erlang behaviours" do
     assert get_callbacks_with_docs(:supervisor) == [%{
@@ -116,21 +146,11 @@ defmodule ElixirSense.Core.IntrospectionTest do
     ]
   end
 
-  test "get_types ignores privates types (:opaque and :typep)" do
-    types = get_types(ModuleWithPrivateTypes)
-    assert types == [type: {:type_t, {:type, 4, :atom, []}, []}]
-  end
-
-  test "get_func_docs_md works for kernel special forms" do
-    docs = get_func_docs_md(Kernel.SpecialForms, :__MODULE__)
-    assert docs == "> Kernel.SpecialForms.__MODULE__()\n\nReturns the current module name as an atom or `nil` otherwise.\n\nAlthough the module can be accessed in the `__ENV__/0`, this macro\nis a convenient shortcut.\n"
-  end
-
   defp get_type_ast(module, type) do
     {_kind, type} =
-      get_types(module)
+      NormalizedTypespec.get_types(module)
       |> Enum.find(fn {_, {name, _, _}} -> name == type end)
-    type_to_quoted(type)
+      NormalizedTypespec.type_to_quoted(type)
   end
 
 end
