+++ test/elixir_sense/core/normalized_typespec_test.exs
@@ -1,0 +1,11 @@
+defmodule ElixirSense.Core.NormalizedTypespecTest do
+
+  use ExUnit.Case
+
+  import ElixirSense.Core.Normalized.Typespec
+
+  test "get_types ignores privates types (:opaque and :typep)" do
+    types = get_types(ModuleWithPrivateTypes)
+    assert types == [type: {:type_t, {:type, 4, :atom, []}, []}]
+  end
+end
