+++ test/support/module_wiht_private_types.ex
@@ -2,4 +2,9 @@ defmodule ModuleWithPrivateTypes do
   @opaque opaque_t :: atom
   @typep typep_t :: atom
   @type type_t :: atom
+
+  @spec just_to_use_typep(typep_t) :: typep_t
+  def just_to_use_typep(t) do
+    t
+  end
 end
